refactor: about and remote video dialogs
