+++ src/layouts/MainLayout.vue
@@ -188,129 +188,11 @@
                 </q-date>
               </q-popup-proxy>
             </q-btn>
-            <q-dialog v-model="remoteVideoPopup">
-              <div
-                class="items-center q-pb-lg q-px-lg q-gutter-y-lg bg-secondary-contrast large-overlay"
-              >
-                <div class="text-h6 row">{{ $t('add-video-jw-org') }}</div>
-                <div class="row">{{ $t('add-a-video-explain') }}</div>
-                <!-- <div class="row">
-                  <q-linear-progress
-                    :value="remoteVideosLoadingProgress"
-                    class="q-mt-md"
-                  />
-                </div> -->
-                <div class="row">
-                  <div class="col-grow">
-                    <q-input
-                      :label="$t('search')"
-                      clearable
-                      dense
-                      outlined
-                      v-model="remoteVideoFilter"
-                    >
-                      <template v-slot:prepend>
-                        <q-icon name="mmm-search" />
-                      </template>
-                    </q-input>
-                  </div>
-                </div>
-                <div
-                  class="text-center row items-center justify-center"
-                  v-if="
-                    !!(
-                      remoteVideosLoadingProgress < 1 &&
-                      (remoteVideosFiltered.length === 0 || remoteVideoFilter)
-                    )
-                  "
-                >
-                  <q-spinner color="primary" size="md" />
-                </div>
-                <div class="row">
-                  <q-scroll-area
-                    :bar-style="barStyle()"
-                    :thumb-style="thumbStyle()"
-                    style="width: 100vw; height: 40vh"
-                  >
-                    <div class="row q-col-gutter-md">
-                      <template
-                        :key="video.guid"
-                        v-for="video in remoteVideosFiltered"
-                      >
-                        <div
-                          class="col-xs-6 col-sm-4 col-md-3 col-lg-3 col-xl-2"
-                        >
-                          <div
-                            :class="{
-                              'cursor-pointer': true,
-                              'rounded-borders-lg': true,
-                              'full-height': true,
-                              'bg-accent-100':
-                                hoveredRemoteVideo === video.guid,
-                            }"
-                            @click="
-                              downloadAdditionalRemoteVideo(
-                                video.files,
-                                getBestImageUrl(video.images, 'md'),
-                              );
-                              remoteVideoPopup = false;
-                            "
-                            @mouseout="hoveredRemoteVideo = ''"
-                            @mouseover="hoveredRemoteVideo = video.guid"
-                            flat
-                            v-ripple
-                          >
-                            <q-card-section class="q-pa-sm">
-                              <q-img
-                                :src="getBestImageUrl(video.images, 'md')"
-                                class="rounded-borders"
-                              >
-                                <q-badge
-                                  class="q-mt-sm q-ml-sm bg-semi-black rounded-borders-sm"
-                                  style="padding: 5px !important"
-                                >
-                                  <q-icon
-                                    class="q-mr-xs"
-                                    color="white"
-                                    name="mmm-play"
-                                  />
-                                  {{ formatTime(video.duration) }}
-                                </q-badge>
-                              </q-img>
-                            </q-card-section>
-                            <q-card-section class="q-pa-sm">
-                              <div class="text-subtitle2 q-mb-xs">
-                                {{ video.title }}
-                              </div>
-                              <div>
-                                <span class="text-caption text-dark-grey">{{
-                                  video.naturalKey
-                                }}</span>
-                              </div>
-                            </q-card-section>
-                          </div>
-                        </div>
-                      </template>
-                    </div>
-                  </q-scroll-area>
-                </div>
-                <div class="row items-center">
-                  <div class="col">
-                    <q-toggle
-                      :label="$t('include-audio-description')"
-                      checked-icon="mmm-check"
-                      color="primary"
-                      v-model="remoteVideosIncludeAudioDescription"
-                    />
-                  </div>
-                  <div class="col text-right">
-                    <q-btn color="negative" flat v-close-popup>{{
-                      $t('cancel')
-                    }}</q-btn>
-                  </div>
-                </div>
-              </div>
-            </q-dialog>
+            <DialogRemoteVideo
+              :remote-videos="remoteVideos"
+              :remote-videos-loading-progress="remoteVideosLoadingProgress"
+              v-model="remoteVideoPopup"
+            />
           </template>
           <template v-else-if="route.fullPath === '/settings'">
             <q-btn color="white-transparent" unelevated v-if="selectedDate">
@@ -637,86 +519,7 @@
         </q-card-actions>
       </q-card>
     </q-dialog>
-    <q-dialog v-model="aboutModal">
-      <div
-        class="items-center q-pb-lg q-px-lg q-gutter-y-lg bg-secondary-contrast"
-      >
-        <div class="row items-center">
-          <div class="col-shrink q-mr-md">
-            <img
-              col
-              src="favicon.ico"
-              style="max-height: 10vh; display: flex"
-            />
-          </div>
-          <div class="col">
-            <div class="row text-h6">
-              {{ $t('meeting-media-manager') }}
-            </div>
-            <div class="row">v{{ appVersion }}</div>
-          </div>
-        </div>
-        <div class="row">
-          <div class="col">
-            {{ $t('app-description') }}
-          </div>
-        </div>
-        <div class="row q-gutter-x-md">
-          <div class="col">
-            <q-btn
-              @click="openExternalWebsite(githubLink)"
-              class="q-pa-md full-width"
-              color="accent-400"
-              no-caps
-              outline
-              target="_blank"
-            >
-              <div class="row q-gutter-x-md full-width items-center">
-                <div class="col-shrink text-primary q-ml-none">
-                  <q-icon name="mmm-github" />
-                </div>
-                <div class="col-shrink text-secondary">
-                  {{ $t('github-repo') }}
-                </div>
-                <div class="col text-right text-accent-400">
-                  <q-icon name="mmm-arrow-outward" />
-                </div>
-              </div>
-            </q-btn>
-          </div>
-          <div class="col">
-            <q-btn
-              @click="openExternalWebsite(docsLink)"
-              class="q-pa-md full-width"
-              color="accent-400"
-              no-caps
-              outline
-              target="_blank"
-            >
-              <div class="row q-gutter-x-md full-width items-center">
-                <div class="col-shrink text-primary q-ml-none">
-                  <q-icon name="mmm-guide" />
-                </div>
-                <div class="col-shrink text-secondary">
-                  {{ $t('user-guide') }}
-                </div>
-                <div class="col text-right text-accent-400">
-                  <q-icon name="mmm-arrow-outward" />
-                </div>
-              </div>
-            </q-btn>
-          </div>
-        </div>
-        <div class="row">
-          <div class="col">
-            {{ $t('app-issues') }}
-          </div>
-        </div>
-        <div class="row justify-end">
-          <q-btn flat v-close-popup>{{ $t('close') }}</q-btn>
-        </div>
-      </div>
-    </q-dialog>
+    <DialogAbout v-model="aboutModal" />
   </q-layout>
 </template>
 
@@ -730,6 +533,8 @@ import { get } from 'src/boot/axios';
 import { queues } from 'src/boot/globals';
 import { barStyle, thumbStyle } from 'src/boot/globals';
 import { refreshDateLocale } from 'src/boot/i18n';
+import DialogAbout from 'src/components/dialog/DialogAbout.vue';
+import DialogRemoteVideo from 'src/components/dialog/DialogRemoteVideo.vue';
 import PublicTalkMediaPicker from 'src/components/media/PublicTalkMediaPicker.vue';
 import SongPicker from 'src/components/media/SongPicker.vue';
 import ActionIsland from 'src/components/ui/ActionIsland.vue';
@@ -748,16 +553,11 @@ import {
   getTempDirectory,
   removeEmptyDirs,
 } from 'src/helpers/fs';
-import {
-  downloadAdditionalRemoteVideo,
-  downloadSongbookVideos,
-  getBestImageUrl,
-} from 'src/helpers/jw-media';
+import { downloadSongbookVideos } from 'src/helpers/jw-media';
 import {
   registerAllCustomShortcuts,
   unregisterAllCustomShortcuts,
 } from 'src/helpers/keyboardShortcuts';
-import { formatTime } from 'src/helpers/mediaPlayback';
 import { createTemporaryNotification } from 'src/helpers/notifications';
 import { useAppSettingsStore } from 'src/stores/app-settings';
 import { useCongregationSettingsStore } from 'src/stores/congregation-settings';
@@ -854,12 +654,10 @@ const {
   closeWebsiteWindow,
   fs,
   getAppDataPath,
-  getAppVersion,
   getUserDataPath,
   getUserDesktopPath,
   klawSync,
   navigateWebsiteWindow,
-  openExternalWebsite,
   openWebsiteWindow,
   path,
   pathToFileURL,
@@ -1081,9 +879,6 @@ const datePickerActive = ref(false);
 const remoteVideoPopup = ref(false);
 const remoteVideosLoadingProgress = ref(0);
 const remoteVideos: Ref<MediaItemsMediatorItem[]> = ref([]);
-const remoteVideoFilter = ref('');
-const remoteVideosIncludeAudioDescription = ref(false);
-const hoveredRemoteVideo = ref('');
 const moreOptionsMenuActive = ref(false);
 const cacheFiles: Ref<CacheFile[]> = ref([]);
 const calculatingCacheSize = ref(false);
@@ -1437,21 +1232,6 @@ const getJwVideos = async () => {
   }
 };
 
-const remoteVideosFiltered = computed(() => {
-  const useableVideos = ref(
-    remoteVideos.value.filter(
-      (v) =>
-        remoteVideosIncludeAudioDescription.value ||
-        !v.primaryCategory.endsWith('AD'),
-    ),
-  );
-  if (remoteVideoFilter.value?.length > 2)
-    useableVideos.value = useableVideos.value.filter((video) =>
-      video.title.toLowerCase().includes(remoteVideoFilter.value.toLowerCase()),
-    );
-  return useableVideos.value.slice(0, 50);
-});
-
 const getEventDayColor = (eventDate: string) => {
   try {
     if (!lookupPeriod.value || !currentCongregation.value)
@@ -1484,9 +1264,6 @@ const createNewCongregation = () => {
 };
 
 const aboutModal = ref(false);
-const appVersion = getAppVersion();
-const githubLink = 'https://github.com/sircharlo/mmm-refactor';
-const docsLink = 'https://sircharlo.github.io/mmm-refactor/';
 
 const importMenu: Ref<QMenu | undefined> = ref();
 const openImportMenu = () => {
